B 修正flex架构 构建bison测试体系和编写基本框架 初步实现错误抛出到syntax_error等待后续处理

+++ src/token/token.cpp
@@ -1,4 +1,5 @@
 #include "token.hpp"
+#include <string>
 /*
 *   @brief Token类，用于表示词法分析的结果
 */
@@ -52,8 +53,8 @@ Token::Token(){};
     *   @brief 获取token的int类型值
     *   @return token的int类型值
     */
-    int Token::get_int(){
-        return std::stoi(this->token_value);
+    long long Token::get_int(){
+        return std::stoll(this->token_value);
     };
     /*
     *   @brief 获取token的double类型值
